(refactor) move useMakeReviewCard functionality to useReview and useReviewState
