+++ frontend/src/components/review/hooks/useReview.tsx
@@ -3,8 +3,8 @@ import { Term, TermUpdateObject } from "gql/codegen-output";
 import { useCreateReviewSessionMutation } from "gql/hooks/reviewSession-query";
 import { makeNewSaturationLevels } from "helpers/srs/saturation";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import ReviewCard from "../sub/ReviewCard";
 import { useInitializeReview } from "./useInitializeReview";
-import { useMakeReviewCard } from "./useMakeReviewCard";
 import { useReviewState } from "./useReviewState";
 
 type TermUpdatePassfail = {
@@ -33,7 +33,6 @@ export function useReview() {
 
 	const startedRef = useRef<boolean>();
 
-	const { makeReviewCard, backWasShown, setBackWasShown } = useMakeReviewCard();
 	const {
 		reviewSettings,
 		setReviewSettings,
@@ -44,7 +43,22 @@ export function useReview() {
 		termUpdateArray,
 		setTermUpdateArray,
 		newReviewSession,
+		backWasShown,
+		setBackWasShown,
 	} = useReviewState();
+
+	const makeReviewCard = useCallback(
+		(term: Term) => {
+			return (
+				<ReviewCard
+					direction={reviewSettings.direction}
+					term={term}
+					setBackWasShown={setBackWasShown}
+				/>
+			);
+		},
+		[reviewSettings.direction, setBackWasShown]
+	);
 
 	const { mutate: mutateCreateReviewSession, data: mutateResponse } =
 		useCreateReviewSessionMutation(() => setReviewStage("after"));
